fix: change received field to integer
In sample format v1 and legacy android profiles, `received` is either
expected to be either a Time "parsable" (rfc 3339 formatted string) or
an int.

Therefore we're changing the type to int in V1 implementation to make it
compatible with what vroom expect.

+++ src/profile.rs
@@ -173,8 +173,8 @@ impl Profile {
     /// Returns the received timestamp.
     ///
     /// Returns:
-    ///     float
-    fn get_received(&self) -> f64 {
+    ///     int
+    fn get_received(&self) -> i64 {
         self.profile.get_received()
     }
